backend/requirements.txt
fastapi>=0.111.0,<1.0
uvicorn[standard]>=0.29.0,<1.0
sqlalchemy>=2.0.36,<3.0
python-telegram-bot>=21.6,<22.0
google-genai>=1.0
python-dotenv>=1.0.1
pillow>=10.4.0
httpx>=0.27.0,<1.0
pydantic>=2.10.0,<3.0
aiosqlite>=0.20.0
python-multipart>=0.0.9